Updated contracts and code refactor.
