+++ .env.template
@@ -1,0 +1,4 @@
+HR_EXPRESS_PORT=
+HR_MDB_HOST=
+HR_MDB_PORT=
+HR_MDB_DB=
