+++ src/config/dbConfigs.ts
@@ -1,0 +1,6 @@
+export const HR_MDB_HOST = process.env.HR_MDB_HOST;
+export const HR_MDB_PORT = process.env.HR_MDB_PORT;
+export const HR_MDB_DB = process.env.HR_MDB_DB;
+
+export const HR_MDB_USER = process.env.HR_MDB_USER;
+export const HR_MDB_PASSWORD = process.env.HR_MDB_PASSWORD;
