+++ src/config/expressConfigs.ts
@@ -1,0 +1,1 @@
+export const HR_EXPRESS_PORT = process.env.HR_EXPRESS_PORT;
