+++ src/config/index.ts
@@ -2,11 +2,5 @@ import dotenv from "dotenv";
 
 dotenv.config();
 
-export const HR_EXPRESS_PORT = process.env.HR_EXPRESS_PORT;
-
-export const HR_MDB_HOST = process.env.HR_MDB_HOST;
-export const HR_MDB_PORT = process.env.HR_MDB_PORT;
-export const HR_MDB_DB = process.env.HR_MDB_DB;
-
-export const HR_MDB_USER = process.env.HR_MDB_USER;
-export const HR_MDB_PASSWORD = process.env.HR_MDB_PASSWORD;
+export * from "./expressConfigs";
+export * from "./dbConfigs";
