+++ src/db/horse/schema.ts
@@ -4,5 +4,5 @@ const { Schema } = mongoose;
 export const horseSchema = new Schema({
   name: { type: String, required: true },
   date: { type: Date, default: Date.now },
-  isDeleted: { type: Boolean, default: false },
+  is_deleted: { type: Boolean, default: false },
 });
