+++ src/server/index.ts
@@ -1,9 +1,1 @@
-import express from "express";
-import * as bodyParser from "body-parser";
-
-const server = express();
-
-server.use(bodyParser.urlencoded({ extended: false }));
-server.use(bodyParser.json());
-
-export { server };
+export * from "./serverWithMiddlewares";
