+++ src/server/serverWithMiddlewares.ts
@@ -1,0 +1,9 @@
+import express from "express";
+import * as bodyParser from "body-parser";
+
+const server = express();
+
+server.use(bodyParser.urlencoded({ extended: false }));
+server.use(bodyParser.json());
+
+export { server };
